space_server: dhcpd: add hosts to .dhcp zone
Only requests with the host-name header set will be added to
the dhcp zone (so it can be looked up at <hostname>.dhcp).
This will fail if the host-name is not a valid dns name though.

Also leases on the member wired, member wireless, and free
are added to the zone, but just like the s zone the
dhcp zone and reverse dns for 10.42.0.0/16 cannot be queried
from Labitat (free). This way you can connect gadgets to
Labitat (free), but still look them up in dns and connect
to them from the member networks.