space_server: bird: fiberby: enable TTL security
This protects us amount otherthings against 3rd parties
resetting the TCP connection underneat our BGP sessions.

This has been enabled in both ends, and this
_MUST_ remain enabled, otherwise these sessions
will go down.

If this needs to be disabled for some reason
then it must be coordinated with Fiberby.

RFC 5082 - The Generalized TTL Security Mechanism
https://datatracker.ietf.org/doc/html/rfc5082