space_server: bird: export prefixes based on communities
We only announce a prefix, if we have recieved it from a customer
connection or if we originate it our self.

This way we avoid announcing prefixes matching the earlier used,
prefix list if we haven't recieved it via the customer connection.
This is important for multi-homed labicolo customers.

Esmil: consistent brace placement